requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
bootstrap4==0.1.0
dj-database-url==0.5.0
Django==3.0.8
django-bootstrap4==2.2.0
django-heroku==0.3.1
django-taggit==1.2.0
gunicorn==20.0.4
psycopg2==2.8.5
pytz==2020.1
soupsieve==2.0.1
sqlparse==0.3.1
whitenoise==5.2.0
